+++ common/src/main/java/com/iamkaf/amber/AmberMod.java
@@ -1,18 +1,19 @@
 package com.iamkaf.amber;
 
 import com.iamkaf.amber.api.core.v2.AmberModInfo;
+import com.iamkaf.amber.api.event.v1.events.common.CommandEvents;
+import com.iamkaf.amber.api.platform.v1.ModInfo;
+import com.iamkaf.amber.api.platform.v1.Platform;
 import com.iamkaf.amber.api.registry.v1.DeferredRegister;
 import com.iamkaf.amber.api.registry.v1.RegistrySupplier;
 import com.iamkaf.amber.platform.Services;
 import com.iamkaf.amber.util.Env;
 import com.iamkaf.amber.util.EnvExecutor;
+import com.mojang.brigadier.Command;
+import net.minecraft.commands.Commands;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.player.Player;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.level.Level;
 
 import java.util.ArrayList;
 
@@ -38,5 +39,25 @@ public class AmberMod {
         EnvExecutor.runInEnv(Env.SERVER, () -> Services.AMBER_EVENT_SETUP::registerServer);
 
         ITEMS.register();
+
+        CommandEvents.EVENT.register((dispatcher, registryAccess, environment) -> {
+            Constants.LOG.info("Registering Amber commands for {}", Services.PLATFORM.getPlatformName());
+            dispatcher.register(Commands.literal("amber").executes(commandContext -> {
+                ModInfo amberInfo = Platform.getModInfo(Constants.MOD_ID);
+
+                // wat??
+                if (amberInfo == null) {
+                    commandContext.getSource().sendFailure(Component.literal("Amber mod info not found!"));
+                    return Command.SINGLE_SUCCESS;
+                }
+
+                commandContext.getSource()
+                        .sendSuccess(
+                                () -> Component.literal(amberInfo.name()).append(" - Version: " + amberInfo.version()),
+                                false
+                        );
+                return Command.SINGLE_SUCCESS;
+            }));
+        });
     }
 }