+++ common/src/main/java/com/iamkaf/amber/api/item/SmartTooltip.java
@@ -99,6 +99,7 @@ public class SmartTooltip {
      * @deprecated Use {@link #into(Consumer)} instead. This is only here so mods that use
      * SmartTooltip can still compile.
      */
+    @Deprecated
     public void into(List<Component> tooltipAdder) {
         tooltipAdder.addAll(tooltipComponents);
     }