+++ forge/src/main/java/com/iamkaf/amber/platform/ForgePlatformHelper.java
@@ -1,14 +1,18 @@
 package com.iamkaf.amber.platform;
 
+import com.iamkaf.amber.api.platform.v1.ModInfo;
 import com.iamkaf.amber.platform.services.IPlatformHelper;
 import com.iamkaf.amber.util.Env;
+import net.minecraftforge.fml.ModContainer;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.loading.FMLLoader;
 import net.minecraftforge.fml.loading.FMLPaths;
 import net.minecraftforge.forgespi.language.IModInfo;
+import org.jetbrains.annotations.Nullable;
 
 import java.nio.file.Path;
 import java.util.Collection;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 public class ForgePlatformHelper implements IPlatformHelper {
@@ -53,5 +57,21 @@ public class ForgePlatformHelper implements IPlatformHelper {
     @Override
     public Collection<String> getModIds() {
         return ModList.get().getMods().stream().map(IModInfo::getModId).collect(Collectors.toList());
+    }
+
+    @Override
+    public @Nullable com.iamkaf.amber.api.platform.v1.ModInfo getModInfo(String modId) {
+        Optional<ModContainer> maybeContainer =
+                ModList.get().getLoadedMods().stream().filter(c -> c.getModId().equals(modId)).findFirst();
+        if (maybeContainer.isPresent()) {
+            ModContainer container = maybeContainer.get();
+            return new ModInfo(
+                    container.getModId(),
+                    container.getModInfo().getDisplayName(),
+                    container.getModInfo().getVersion().toString(),
+                    container.getModInfo().getDescription()
+            );
+        }
+        return null;
     }
 }