+++ forge/src/main/java/com/iamkaf/amber/platform/ForgeRegistrarManager.java
@@ -20,7 +20,7 @@ import java.util.function.Supplier;
 public class ForgeRegistrarManager implements IRegistrarManager {
     private final Map<String, Map<ResourceKey<? extends Registry<?>>, DeferredRegister<?>>> registers = new HashMap<>();
 
-    @SuppressWarnings("unchecked")
+    @SuppressWarnings({"unchecked", "removal"})
     private <T> DeferredRegister<T> getRegister(String modId, ResourceKey<Registry<T>> key) {
         Map<ResourceKey<? extends Registry<?>>, DeferredRegister<?>> map = registers.computeIfAbsent(modId, m -> new HashMap<>());
         return (DeferredRegister<T>) map.computeIfAbsent(key, k -> {